pyproject.toml
[tool.poetry]
name = "andromabot"
version = "1.0.6"
description = "Andromeda Labs discord bot"
authors = ["Starship IBC <shrugs.rest0x@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = {path = "./dependencies/discord.py-2.0.0a4405+ge0341c9e.tar.gz"}
stargazeutils = {path = "./dependencies/stargazeutils-0.2.0.tar.gz"}
PyYAML = "^6.0"
Wand = "^0.6.10"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = "--cov=andromabot --cov-report html"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
